requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
Django==4.0
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-database-url==1.0.3
django-js-asset==2.2.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
